move flashmint code
